change all floats to doubles so comparisons actually work

+++ backend/BIGGYM/api/register/spRegisterTrainingPlan.sql
@@ -28,7 +28,7 @@ create procedure spRegisterTrainingPlan(in vNewOrUpdatable_TrainingPlanName varc
                                         in vPerson_LastName varchar(32),
                                         in vPerson_BirthDate date,
                                         in vPerson_Gender char(1),
-                                        in vPerson_BodyHeight float,
+                                        in vPerson_BodyHeight double,
                                      inout ObjectId mediumint unsigned,
                                        out ReturnCode int,
                                        out ErrorCode int,
